+++ CHANGES.rst
@@ -127,6 +127,9 @@ astropy.utils
 astropy.visualization
 ^^^^^^^^^^^^^^^^^^^^^
 
+- Add support for setting ``set_separator(None)`` in WCSAxes to use default
+  separators. [#7570]
+
 astropy.wcs
 ^^^^^^^^^^^